added modular items for headers and updated the navigation bar
