+++ babel.config.js
@@ -12,7 +12,6 @@ module.exports = function (api) {
             src: "./src",
             components: "./src/components",
             constants: "./src/constants",
-            redux: "./src/redux",
             screens: "./src/screens",
             tabs: "./src/tabs",
           },
