+++ package.json
@@ -21,12 +21,13 @@
     "module-resolver": "^1.0.0",
     "prop-types": "^15.8.1",
     "react": "18.2.0",
-    "react-native-deck-swiper": "^2.0.13",
     "react-dom": "18.2.0",
     "react-native": "0.71.7",
+    "react-native-deck-swiper": "^2.0.13",
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-tab-view": "^3.5.1",
+    "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.18.10",
     "react-navigation": "^4.4.4",
     "react-redux": "^8.0.5",
@@ -42,7 +43,8 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-native": "^4.0.0",
     "eslint-plugin-react-redux": "^4.0.0",
-    "prettier": "2.8.7"
+    "prettier": "2.8.7",
+    "typescript": "^4.9.4"
   },
   "private": true
 }
