+++ src/App.js
@@ -16,29 +16,74 @@ import {
 import registerRootComponent from "expo/build/launch/registerRootComponent";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 
+import Colors from "constants/Colors.js";
+import StackScreenOptions from "components/modular/StackScreenOptions.jsx";
+
 /*~~~~TABS GO HERE~~~~*/
 import ArtistAlley from "./tabs/ArtistAlley.js";
 import Events from "./tabs/Events.js";
-import Messages from "./tabs/Messages.js";
+import Messages from "./tabs/Messages.jsx";
 import Profile from "./tabs/Profile.js";
 import SwipeStack from "./tabs/SwipeStack.js";
 
 /*~~~~SCREENS GO HERE~~~~*/
 import ArtistPage from "./screens/ArtistPage.js";
 import DetailView from "./screens/DetailView.js";
-import ChatPage from "./screens/ChatPage.js";
+import ChatPage from "./screens/ChatPage.jsx";
 
 const Tab = createBottomTabNavigator();
 const Stack = createNativeStackNavigator();
 
 function Home() {
   return (
-    <Tab.Navigator screenOptions={{ headerShown: false }}>
-      <Tab.Screen name="SwipeStack" component={SwipeStack} />
-      <Tab.Screen name="ArtistAlley" component={ArtistAlley} />
-      <Tab.Screen name="Events" component={Events} />
-      <Tab.Screen name="Messages" component={Messages} />
-      <Tab.Screen name="Profile" component={Profile} />
+    <Tab.Navigator
+      screenOptions={{
+        headerShown: false,
+        tabBarStyle: {
+          height: 90,
+          flexDirection: "row",
+          alignItems: "center",
+          backgroundColor: Colors.PRIMARY,
+          paddingTop: 10,
+          shadowRadius: 15,
+          shadowColor: "#000000",
+          shadowOffset: { width: 0, height: 1 },
+          shadowOpacity: 0.2,
+          elevation: 5,
+        },
+      }}
+    >
+      <Tab.Screen
+        key={1}
+        name="Events"
+        component={Events}
+        options={StackScreenOptions("calendar", 40)}
+      />
+
+      <Tab.Screen
+        key={2}
+        name="ArtistAlley"
+        component={ArtistAlley}
+        options={StackScreenOptions("store", 35)}
+      />
+      <Tab.Screen
+        key={3}
+        name="SwipeStack"
+        component={SwipeStack}
+        options={StackScreenOptions("gavel", 42)}
+      />
+      <Tab.Screen
+        key={4}
+        name="Messages"
+        component={Messages}
+        options={StackScreenOptions("comment", 42)}
+      />
+      <Tab.Screen
+        key={5}
+        name="Profile"
+        component={Profile}
+        options={StackScreenOptions("user-circle", 44, "solid")}
+      />
     </Tab.Navigator>
   );
 }
@@ -47,30 +92,30 @@ function App() {
   return (
     <Provider store={store}>
       <NavigationContainer>
-        <SafeAreaView style={{ flex: 1 }}>
-          <Stack.Navigator>
-            <Stack.Screen
-              name="Home"
-              component={Home}
-              options={{ headerShown: false }}
-            />
-            <Stack.Screen
-              name="ArtistPage"
-              component={ArtistPage}
-              options={{ headerShown: false, gestureDirection: "vertical" }}
-            />
-            <Stack.Screen
-              name="DetailView"
-              component={DetailView}
-              options={{ headerShown: false, gestureDirection: "vertical" }}
-            />
-            <Stack.Screen
-              name="ChatPage"
-              component={ChatPage}
-              options={{ headerShown: false }}
-            />
-          </Stack.Navigator>
-        </SafeAreaView>
+        <Stack.Navigator>
+          <Stack.Screen
+            name="Home"
+            component={Home}
+            options={{ headerShown: false }}
+          />
+          <Stack.Screen
+            name="ArtistPage"
+            component={ArtistPage}
+            options={{ headerShown: false, gestureDirection: "vertical" }}
+          />
+          <Stack.Screen
+            name="DetailView"
+            component={DetailView}
+            options={{ headerShown: false, gestureDirection: "vertical" }}
+          />
+          <Stack.Screen
+            name="ChatPage"
+            component={ChatPage}
+            options={{
+              headerShown: false,
+            }}
+          />
+        </Stack.Navigator>
       </NavigationContainer>
     </Provider>
   );
