+++ src/redux/store.js
@@ -1,11 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 
 import exampleReducer from "./exampleReducer";
-import imageSlice from './imageReducer'
+import imageSlice from "./imageReducer";
 
 export const store = configureStore({
   reducer: {
     message: exampleReducer,
-    images: imageSlice,
   },
 });
