+++ src/tabs/Messages.jsx
@@ -4,25 +4,24 @@ import PropTypes from "prop-types";
 import {
   StyleSheet,
   // Button,
-  View,
-  // SafeAreaView,
-  Text,
+  // View,
+  SafeAreaView,
+  // Text,
   // Alert,
 } from "react-native";
 
 import Colors from "constants/Colors.js";
-import MessagesToolbar from "components/Messages/MessagesToolbar.js";
-import MessagesChatList from "components/Messages/MessagesChatList.js";
+import Header from "components/modular/Header.jsx";
+import MessagesToolbar from "components/Messages/MessagesToolbar.jsx";
+import MessagesChatList from "components/Messages/MessagesChatList.jsx";
 
 const Messages = ({ navigation }) => {
   return (
-    <View style={styles.container}>
-      <View style={{ height: 65, borderWidth: 1 }}>
-        <Text>Messages Header</Text>
-      </View>
+    <SafeAreaView style={styles.container}>
+      <Header navigation={navigation} title="Messages" />
       <MessagesToolbar navigation={navigation} />
       <MessagesChatList navigation={navigation} />
-    </View>
+    </SafeAreaView>
   );
 };
